ai-agents/agno_ai_agent/requirements.txt
agno
openai
python-dotenv
duckduckgo-search
pypdf
lancedb
streamlit
groq
sentence-transformers
pandas
tantivy
yfinance
pandas
